Refactor adminsite API into router/service modules
