+++ admin_panel/adminsite/__init__.py
@@ -1,5 +1,5 @@
 """API and utilities for AdminSite constructor."""
 
-from admin_panel.adminsite.api import router
+from admin_panel.adminsite.router import router
 
 __all__ = ["router"]
